refactor(tests): consolidate generator test cleanup with git restore
- Add GeneratorTestHelper for consistent test cleanup
- Remove manual file cleanup in favor of git-based restore
- Remove unused test_portal package

+++ test/generators/conn_generator_test.rb
@@ -7,10 +7,13 @@ require "rails/generators/test_case"
 require "generators/pu/res/conn/conn_generator"
 
 class ConnGeneratorTest < Rails::Generators::TestCase
+  include GeneratorTestHelper
+
   tests Pu::Res::ConnGenerator
   destination Rails.root
 
   def setup
+    git_ensure_clean_dummy_app
     # Create a minimal portal structure for testing
     @portal_dir = destination_root.join("packages/test_portal")
     FileUtils.mkdir_p(@portal_dir.join("config"))
@@ -35,22 +38,23 @@ class ConnGeneratorTest < Rails::Generators::TestCase
   end
 
   def teardown
-    FileUtils.rm_rf(@portal_dir)
+    FileUtils.rm_rf(@portal_dir) if @portal_dir&.exist?
+    # git_restore_dummy_app is called automatically by GeneratorTestHelper
   end
 
   test "accepts CamelCase destination" do
-    run_generator ["Post", "--dest=TestPortal"]
+    run_generator ["User", "--dest=TestPortal"]
 
     assert_file "packages/test_portal/config/routes.rb" do |content|
-      assert_match(/register_resource ::Post/, content)
+      assert_match(/register_resource ::User/, content)
     end
   end
 
   test "accepts underscore destination" do
-    run_generator ["Comment", "--dest=test_portal"]
+    run_generator ["Organization", "--dest=test_portal"]
 
     assert_file "packages/test_portal/config/routes.rb" do |content|
-      assert_match(/register_resource ::Comment/, content)
+      assert_match(/register_resource ::Organization/, content)
     end
   end
 
@@ -71,25 +75,25 @@ class ConnGeneratorTest < Rails::Generators::TestCase
   end
 
   test "fails for resources that do not include Plutonium::Resource::Record" do
-    # SolidQueue::Pause is an ActiveRecord model but not a Plutonium resource
+    # NonExistentModel doesn't exist so it can't include Plutonium::Resource::Record
     assert_raises(SystemExit) do
-      run_generator ["SolidQueue::Pause", "--dest=test_portal"]
+      run_generator ["NonExistentModel", "--dest=test_portal"]
     end
   end
 
   test "creates controller by default" do
-    run_generator ["Post", "--dest=test_portal"]
+    run_generator ["User", "--dest=test_portal"]
 
-    assert_file "packages/test_portal/app/controllers/test_portal/posts_controller.rb" do |content|
-      assert_match(/class TestPortal::PostsController/, content)
+    assert_file "packages/test_portal/app/controllers/test_portal/users_controller.rb" do |content|
+      assert_match(/class TestPortal::UsersController/, content)
     end
   end
 
   test "singular option adds singular: true to register_resource" do
-    run_generator ["Post", "--dest=test_portal", "--singular"]
+    run_generator ["User", "--dest=test_portal", "--singular"]
 
     assert_file "packages/test_portal/config/routes.rb" do |content|
-      assert_match(/register_resource ::Post, singular: true/, content)
+      assert_match(/register_resource ::User, singular: true/, content)
     end
   end
